Fix #5
This changes the behavior of filtering Divs to be similar to that of
Spans: we now add the raw Latex before and after the content of the
Div rather than adding it inline.

This fixes a bug described in the issue and also changes the behavior
in certain small ways. For example, previously in a list the first
bullet or number would not have its color changed (but later items
would), or horizontal rules at the start or end would not have their
color effected. Now everything enclosed in the Div is within the scope
of the color.